+++ datasourceInfosys-portlet-service/src/main/java/de/hska/wi/awp/datasource/infosys/service/Teilnote_feedbackLocalServiceUtil.java
@@ -263,7 +263,7 @@ public class Teilnote_feedbackLocalServiceUtil {
     }
 
     public static de.hska.wi.awp.datasource.infosys.model.Teilnote_feedback findByFeedback_idAndBewertungskriterium_id(
-        long feedback_id, int bewertungskriterium_id) {
+        long feedback_id, long bewertungskriterium_id) {
         return getService()
                    .findByFeedback_idAndBewertungskriterium_id(feedback_id,
             bewertungskriterium_id);
