+++ datasourceInfosys-portlet-service/src/main/java/de/hska/wi/awp/datasource/infosys/service/Teilnote_feedbackLocalServiceWrapper.java
@@ -278,7 +278,7 @@ public class Teilnote_feedbackLocalServiceWrapper
 
     @Override
     public de.hska.wi.awp.datasource.infosys.model.Teilnote_feedback findByFeedback_idAndBewertungskriterium_id(
-        long feedback_id, int bewertungskriterium_id) {
+        long feedback_id, long bewertungskriterium_id) {
         return _teilnote_feedbackLocalService.findByFeedback_idAndBewertungskriterium_id(feedback_id,
             bewertungskriterium_id);
     }
